move rounds system test into the new test folder structure
