+++ test/integration/rounds/rounds.js
@@ -17,12 +17,12 @@
 const async = require('async');
 const elements = require('lisk-elements').default;
 const Promise = require('bluebird');
-const ed = require('../../../../helpers/ed.js');
-const slots = require('../../../../helpers/slots');
-const Bignum = require('../../../../helpers/bignum.js');
-const accountsFixtures = require('../../../fixtures/accounts');
-const randomUtil = require('../../../common/utils/random');
-const queriesHelper = require('../../common/sql/queriesHelper.js');
+const ed = require('../../../helpers/ed.js');
+const slots = require('../../../helpers/slots');
+const Bignum = require('../../../helpers/bignum.js');
+const queriesHelper = require('../../common/integration/sql/queriesHelper.js');
+const accountsFixtures = require('../../fixtures/accounts');
+const randomUtil = require('../../common/utils/random');
 const localCommon = require('./../common');
 
 const constants = global.constants;
